# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Caficultores",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "Cultivo": {
        "type": "object",
        "required": [
          "nombre"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID único del cultivo"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del cultivo"
          },
          "variedad": {
            "type": "string",
            "description": "Variedad del cultivo"
          },
          "hectareas": {
            "type": "number",
            "description": "Número de hectáreas del cultivo",
            "default": 0
          },
          "ubicacion": {
            "type": "string",
            "description": "Ubicación del cultivo"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de última actualización"
          }
        }
      },
      "CultivoInput": {
        "type": "object",
        "required": [
          "nombre"
        ],
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre del cultivo"
          },
          "variedad": {
            "type": "string",
            "description": "Variedad del cultivo"
          },
          "hectareas": {
            "type": "number",
            "description": "Número de hectáreas del cultivo"
          },
          "ubicacion": {
            "type": "string",
            "description": "Ubicación del cultivo"
          }
        }
      },
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensaje de confirmación"
          },
          "cultivo": {
            "$ref": "#/components/schemas/Cultivo"
          }
        }
      }
    }
  },
  "paths": {
    "/api/cultivos": {
      "get": {
        "summary": "Obtener todos los cultivos",
        "tags": [
          "Cultivos"
        ],
        "responses": {
          "200": {
            "description": "Lista de cultivos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cultivo"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "post": {
        "summary": "Registrar un cultivo",
        "tags": [
          "Cultivos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CultivoInput"
              },
              "example": {
                "nombre": "Café Arábica",
                "variedad": "Caturra",
                "hectareas": 2.5,
                "ubicacion": "Zona Norte"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cultivo registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cultivo"
                }
              }
            }
          },
          "400": {
            "description": "Error en los datos enviados"
          }
        }
      }
    },
    "/api/cultivos/{id}": {
      "get": {
        "summary": "Obtener un cultivo por ID",
        "tags": [
          "Cultivos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del cultivo",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cultivo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cultivo"
                }
              }
            }
          },
          "404": {
            "description": "Cultivo no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "put": {
        "summary": "Actualizar un cultivo por ID",
        "tags": [
          "Cultivos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del cultivo",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CultivoInput"
              },
              "example": {
                "nombre": "Café Arábica Actualizado",
                "variedad": "Bourbon",
                "hectareas": 3.0,
                "ubicacion": "Zona Sur"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cultivo actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cultivo"
                }
              }
            }
          },
          "400": {
            "description": "Error en los datos enviados"
          },
          "404": {
            "description": "Cultivo no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un cultivo por ID",
        "tags": [
          "Cultivos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del cultivo",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cultivo eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Cultivo no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      }
    }
  }
}